# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions": {
      "post": {
        "operationId": "SessionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Session name already exists"
          }
        },
        "summary": "Create a new WhatsApp session",
        "tags": [
          "sessions"
        ]
      },
      "get": {
        "operationId": "SessionController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max sessions to return (1-1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of sessions to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "operationId": "SessionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get session by ID",
        "tags": [
          "sessions"
        ]
      },
      "delete": {
        "operationId": "SessionController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Delete a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/start": {
      "post": {
        "operationId": "SessionController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session already started"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Start a session and initialize WhatsApp connection",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/stop": {
      "post": {
        "operationId": "SessionController_stop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session stopped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Stop a session and disconnect WhatsApp",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/force-kill": {
      "post": {
        "operationId": "SessionController_forceKill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session force-killed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Force-kill a stuck session (SIGKILL its wedged engine, then tear it down)",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/qr": {
      "get": {
        "operationId": "SessionController_getQRCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QR code data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QRCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "QR code not ready or session already authenticated"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get QR code for session authentication",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/pairing-code": {
      "post": {
        "operationId": "SessionController_requestPairingCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPairingCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing code generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not started or already authenticated"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Request an 8-char pairing code to link via phone number (alternative to QR)",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/groups": {
      "get": {
        "operationId": "SessionController_getGroups",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max groups to return (1–1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of groups to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of groups the session is a member of"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get all groups for a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats": {
      "get": {
        "operationId": "SessionController_getChats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max chats to return (1–1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of chats to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of active chats (most recent first)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get active chats for a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/read": {
      "post": {
        "operationId": "SessionController_markChatRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkChatReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat marked as read successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Mark a chat as read/seen",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/unread": {
      "post": {
        "operationId": "SessionController_markChatUnread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkChatReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat marked as unread successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Mark a chat as unread",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/delete": {
      "post": {
        "operationId": "SessionController_deleteChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteChatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat deleted successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Delete a chat from the chat list (e.g. a group you have left)",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/sessions/{sessionId}/groups": {
      "post": {
        "operationId": "GroupController_create",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created"
          }
        },
        "summary": "Create a new group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/stats/sessions/{sessionId}": {
      "get": {
        "operationId": "StatsController_getSessionStats",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-session statistics for the requested session."
          }
        },
        "summary": "Get statistics for a specific session",
        "tags": [
          "statistics"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.10.8",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique name for the session (alphanumeric and hyphens only)",
            "example": "my-bot",
            "minLength": 3,
            "maxLength": 50
          },
          "config": {
            "type": "object",
            "description": "Session configuration options. Set autoRejectCalls (boolean, default false) to automatically reject incoming calls — the call.received event is still emitted.",
            "example": {
              "autoReconnect": true
            }
          },
          "proxyUrl": {
            "type": "string",
            "description": "Optional per-session egress proxy URL (http/https/socks4/socks5; credentialed form \"http://user:pass@host\" allowed). Must be a REAL, REACHABLE proxy — an unreachable value silently blocks the WhatsApp WebSocket (no QR is ever delivered) and the session start times out (~30s → 504 Gateway Timeout). Leave unset unless your network cannot reach WhatsApp directly."
          },
          "proxyType": {
            "type": "string",
            "description": "Proxy type",
            "enum": [
              "http",
              "https",
              "socks4",
              "socks5"
            ],
            "example": "http"
          }
        },
        "required": [
          "name"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "sess_123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "my-bot"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "initializing",
              "qr_ready",
              "authenticating",
              "ready",
              "disconnected",
              "failed"
            ],
            "example": "ready"
          },
          "phone": {
            "type": "string",
            "example": "628123456789",
            "nullable": true
          },
          "pushName": {
            "type": "string",
            "example": "John Doe",
            "nullable": true
          },
          "connectedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-02T10:00:00Z",
            "nullable": true
          },
          "lastActive": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-02T10:30:00Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-02-02T09:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-02-02T10:00:00Z"
          },
          "lastError": {
            "type": "string",
            "description": "Human-readable reason for the most recent terminal engine failure (only set when status is FAILED).",
            "example": "Failed to launch the browser process: spawn /usr/bin/chromium ENOENT",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "QRCodeResponseDto": {
        "type": "object",
        "properties": {
          "qrCode": {
            "type": "string",
            "description": "QR code as data URL",
            "example": "data:image/png;base64,..."
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "initializing",
              "qr_ready",
              "authenticating",
              "ready",
              "disconnected",
              "failed"
            ],
            "example": "qr_ready"
          }
        },
        "required": [
          "qrCode",
          "status"
        ]
      },
      "RequestPairingCodeDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number to link, digits only in international format (country code + number).",
            "example": "628123456789"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "PairingCodeResponseDto": {
        "type": "object",
        "properties": {
          "pairingCode": {
            "type": "string",
            "description": "The 8-character pairing code to enter in WhatsApp.",
            "example": "ABCD1234"
          },
          "status": {
            "type": "string",
            "description": "Current session status.",
            "example": "qr_ready"
          }
        },
        "required": [
          "pairingCode",
          "status"
        ]
      },
      "ChatSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "628111@c.us"
          },
          "name": {
            "type": "string",
            "example": "Alice"
          },
          "isGroup": {
            "type": "boolean",
            "description": "Retained for back-compat; true for @g.us chats.",
            "example": false
          },
          "kind": {
            "type": "string",
            "enum": [
              "individual",
              "group",
              "channel",
              "status",
              "broadcast",
              "unknown"
            ],
            "description": "User-facing chat kind.",
            "example": "individual"
          },
          "unreadCount": {
            "type": "number",
            "example": 1
          },
          "timestamp": {
            "type": "number",
            "description": "Unix seconds of the last activity.",
            "example": 1700000010
          },
          "lastMessage": {
            "type": "string",
            "example": "hi"
          }
        },
        "required": [
          "id",
          "name",
          "isGroup",
          "kind",
          "unreadCount",
          "timestamp"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}